requirements.txt
flask_sqlalchemy
flask_login
flask_sqlalchemy
flask_wtf
flask_bcrypt